fix: cast EPOLLET to int to satisfy -Wpedantic
EPOLLET is 0x80000000u which overflows a signed int enumerator.
Cast preserves the bit pattern through epoll's uint32_t round-trip.

+++ core/poll.h
@@ -181,7 +181,7 @@ typedef enum poll_events_t {
     POLL_EVENT_ERROR = EPOLLERR,
     POLL_EVENT_HUP = EPOLLHUP,
     POLL_EVENT_RDHUP = EPOLLRDHUP,
-    POLL_EVENT_EDGE = EPOLLET,
+    POLL_EVENT_EDGE = (int)EPOLLET,
 } poll_events_t;
 #elif defined(OS_MACOS)
 #include <sys/event.h>
